feat/improved-import-time-by-external-separation
