+++ pyproject.toml
@@ -43,6 +43,7 @@ dev = [
     "pytest>=8.3.3",
     "seaborn>=0.13.2",
     "treescope>=0.1.6",
+    "tune>=0.1.6",
 ]
 
 [tool.pyright]
